fix: correct typos
