+++ .gitignore
@@ -1,3 +1,6 @@
+# Private files
+*.xlsx
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
