--- ultis.py
@@ -1,24 +1,0 @@
-import tkinter
-from tkinter import filedialog
-
-
-class Features:
-    def __init__(self, interface):
-        self.interface = interface
-        self.interface.filename = None
-
-    def get_files(self):
-        self.interface.filename = filedialog.askopenfilename(initialdir='/',
-                                                             filetypes=(('excel files', '*.xlsx'),
-                                                                        ('all files', '*.*')), )
-
-    def finish(self):
-        self.interface.destroy()
-
-    def put_in(self, strVar):
-        return tkinter.Entry(self.interface, textvariable=strVar, show=None, font=('Arial', 14))
-
-    def print_filename(self):
-        return self.interface.filename
-
-
